refactor(ocale): migrate places update wire code into domains/cloud-v2
Moves the cloud-v2 update Operation wire code (builder, parser, wire
types, rate-limit constant, parsed shapes) for the Places Resource into
a new per-shape sub-tree under src/domains/cloud-v2/places/, following
ADR-022. resources/places/client.ts now imports the update bits from
the Domain sub-tree; the legacy /universes/v1/.../versions publish/save
Operation continues to live in resources/places/ until its own Domain
folder is established. Public surface, wire requests, parsed shapes,
and the places.update operation key are preserved end-to-end.
