+++ packages/open-cloud/src/resources/places/operations.ts
@@ -1,8 +1,5 @@
 import type { OperationLimit } from "../../internal/http/rate-limit-queue.ts";
 
-const UPDATE_PER_MINUTE = 100;
-const SECONDS_PER_MINUTE = 60;
-
 /**
  * Per-second request ceiling for publishing or saving a place version,
  * from the Open Cloud OpenAPI schema (30 requests per minute, expressed
@@ -17,32 +14,10 @@ export const PUBLISH_OPERATION_LIMIT: OperationLimit = Object.freeze({
 });
 
 /**
- * Per-second request ceiling for updating a place, from the Open Cloud
- * OpenAPI schema (100 requests per minute per API key owner). Keyed
- * independently from {@link PUBLISH_OPERATION_LIMIT} so publish and
- * update do not share a queue; upstream quota accounting is not
- * documented as shared and the conservative default is fewer
- * cross-method contention surprises.
- */
-export const UPDATE_OPERATION_LIMIT: OperationLimit = Object.freeze({
-	maxPerSecond: UPDATE_PER_MINUTE / SECONDS_PER_MINUTE,
-	operationKey: "places.update",
-});
-
-/**
  * Scopes required to publish or save a place version, sourced from
  * `x-roblox-scopes` on the `Places_CreatePlaceVersionApiKey` operation
  * in the vendored OpenAPI schema.
  */
 export const PUBLISH_REQUIRED_SCOPES: ReadonlyArray<string> = Object.freeze([
 	"universe-places:write",
-]);
-
-/**
- * Scopes required to update a place's metadata, sourced from
- * `x-roblox-scopes` on the `Cloud_UpdatePlace` operation in the vendored
- * OpenAPI schema.
- */
-export const UPDATE_REQUIRED_SCOPES: ReadonlyArray<string> = Object.freeze([
-	"universe.place:write",
 ]);
